requirements.txt
Flask
flask-cors
pytesseract
pdfplumber
Pillow
